build: Drop redundant platforms lines in create-package
There are a number of entries in create-package that have
platforms of the form:

    platforms: ["linux-x64", "linux-arm64", "darwin-arm64", "win32-x64"],

Which means, these entries apply to all supported platforms.

It's more concise, however, to just omit these lines entirely,
since lacking a platforms line, the default behavior is to
operate on all platforms.

This commit drops the superfluous platforms filters.